Extractors/OrcusRat/requirements.txt
dnfile==0.8.0
malduck==4.2.0
pycryptodome==3.14.1
